+++ aquadoggo/src/network/replication/mod.rs
@@ -1,0 +1,7 @@
+mod behaviour;
+mod handler;
+mod protocol;
+
+pub use behaviour::Behaviour;
+pub use handler::Handler;
+pub use protocol::{Message, PROTOCOL_NAME};
