+++ aquadoggo/src/network/replication/protocol.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+pub const PROTOCOL_NAME: &[u8] = b"/p2p/p2panda/1.0.0";
+
+pub enum Message {}
